+++ src/interfaces/ITokenLockUp.sol
@@ -16,24 +16,33 @@ interface ITokenLockUp {
 
     event Withdrawal(address indexed _addr, uint256 timestamp, uint256 amount);
 
+    /// @notice Pauses the contract.
     function pause() external;
 
+    /// @notice Unpauses the contract if the deposit deadline has not passed.
     function unpause() external;
 
-    /// @notice Let users withdraw their token
+    /// @notice Allows a user to withdraw their locked tokens if the lock duration has passed.
     /// @param _index key for users struct.
     function withdrawTokens(uint256 _index) external;
 
-    function setDepositDeadline(uint _depositDeadline) external;
+    /// @notice Allows the owner to set the deposit deadline.
+    /// @param _depositDeadline new nft address.
+    function setDepositDeadline(uint256 _depositDeadline) external;
 
+    /// @notice Allows the owner to set the address of the token contract.
+    /// @param _newTokenAddress new token address.
     function setTokenAddress(address _newTokenAddress) external;
 
+    /// @notice Allows the owner to set the address of the NFT contract.
+    /// @param _nftAddress new nft address.
     function setNftContractnAddress(address _nftAddress) external;
 
-    /// @notice Let users deposit their token and recieve NFT(s)
+    /// @notice Allows a user to lock up their tokens for a specified duration in exchange for NFTs.
     /// @param _amount amount to deposit.
     /// @param _lockDuration lock up time.
-    function deposit(uint256 _amount, uint _lockDuration) external;
+    function deposit(uint256 _amount, uint256 _lockDuration) external;
 
-    function getUserLockedBatchTokenCount() external view returns (uint);
+    /// @notice Allows the user to get the count of their locked token batches.
+    function getUserLockedBatchTokenCount() external view returns (uint256);
 }
